+++ src/components/custom-field-input-single-choice/custom-field-input-single-choice.jsx
@@ -12,7 +12,7 @@ import styles from './custom-field-input-single-choice.css';
 import Listbox from '../listbox/listbox.jsx';
 import ListOption from '../list-option/list-option.jsx';
 import NoOptions from '../__internal__/no-options/no-options.jsx';
-import Popup from '../__internal__/popup/popup.jsx';
+import Popup from '../popup/popup.jsx';
 import useValidation from '../../hooks/use-validation.jsx';
 import useDropdownClose from '../../hooks/use-dropdown-close.js';