+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
+  - Minor: `CustomFieldInputSingleChoice` and `CustomFieldInputMultipleChoice` now indicate to the user when no options are available
 </details>
 
 ## 0.31.0 (July 20, 2020)